requirements-freeze.txt
aerich==0.2.1
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
aiosqlite==0.16.1
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
asyncclick==8.1.7.2
asyncpg==0.30.0
attrs==24.3.0
blinker==1.9.0
click==8.1.8
Flask==3.1.0
frozenlist==1.5.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.10
iso8601==0.1.16
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
multidict==6.1.0
passlib==1.7.4
priority==2.0.0
propcache==0.2.1
pydantic==2.10.4
pydantic_core==2.27.2
pyhumps==3.8.0
PyJWT==2.10.1
pypika-tortoise==0.1.6
python-dotenv==0.19.2
pytz==2024.2
Quart==0.20.0
quart-schema==0.21.0
sniffio==1.3.1
toml==0.10.2
tortoise-orm==0.17.4
typing_extensions==4.12.2
uvicorn==0.15.0
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.18.3
